requirements.txt
streamlit==1.37.0
pandas==2.2.2
duckdb==1.1.1
matplotlib==3.9.2
python-dotenv==1.0.1
huggingface_hub==0.24.6
requests==2.32.3

